backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
supabase>=2.10.0
PyJWT[crypto]>=2.9.0
google-genai>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
